Fix removal from BinaryTree
Fixes #1.

When removing a node n, n should be replaced by the minimum value in the right
subtree of n, or the maximum value in the left subtree of n. The current
implementation incorrectly replaced the n with the maximum value in the
right subtree of n. This fixes the issue by using the minimum value in
the right subtree of n instead.